# packages/shared-contract/openapi/agentum.openapi.yaml
openapi: 3.1.0
info:
  title: Agentum API
  version: 0.1.0
  description: Agentum 平台 API 契约占位文件。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
paths:
...
  /api/admin/tenants/{tenantId}/organization/overview:
    get:
      summary: 获取租户人员组织概览
      description: 租户管理页使用的租户人员、部门、角色和成员关系聚合视图。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 租户人员组织概览
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrganizationOverview"
        "401":
          description: 未登录或令牌无效
        "403":
          description: 当前账号不能管理该租户
        "404":
          description: 租户不存在或已停用
  /api/admin/tenants/{tenantId}/organization/members:
    post:
      summary: 新增租户成员
      description: 创建用户账号并分配当前租户内部门、角色和空间。后端会校验租户、部门、角色归属。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberRequest"
      responses:
        "200":
          description: 新增成功并返回最新人员组织概览
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrganizationOverview"
        "400":
          description: 请求参数不完整或部门、角色不可用
        "401":
          description: 未登录或令牌无效
        "403":
          description: 当前账号不能管理该租户
        "409":
          description: 用户名已存在
  /api/admin/tenants/{tenantId}/organization/departments:
    post:
      summary: 新增租户部门
      description: 创建当前租户内部门，可选上级部门。部门编码由后端自动生成，后端会校验上级部门归属。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
      responses:
        "200":
          description: 新增成功并返回最新人员组织概览
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrganizationOverview"
        "400":
          description: 请求参数不完整或上级部门不可用
        "401":
          description: 未登录或令牌无效
        "403":
          description: 当前账号不能管理该租户
  /api/admin/tenants/{tenantId}/organization/departments/{departmentId}:
    patch:
      summary: 更新租户部门
      description: 更新当前租户内部门名称、上级部门和排序，部门编码不可由前端修改。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentRequest"
      responses:
        "200":
          description: 更新成功并返回最新人员组织概览
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrganizationOverview"
        "400":
          description: 上级部门不可用或循环引用
        "404":
          description: 部门不存在
    delete:
      summary: 停用租户部门
      description: 仅当部门下没有启用成员时允许停用。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 停用成功
        "400":
          description: 部门下仍有启用成员
        "404":
          description: 部门不存在
  /api/admin/tenants/{tenantId}/organization/roles:
    post:
      summary: 新增租户业务角色
      description: 创建可分配给成员和资源授权的租户内角色，角色编码由后端自动生成。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRoleRequest"
      responses:
        "200":
          description: 新增成功并返回最新人员组织概览
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrganizationOverview"
  /api/admin/tenants/{tenantId}/organization/roles/{roleId}:
    patch:
      summary: 更新租户业务角色
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantRoleRequest"
      responses:
        "200":
          description: 更新成功并返回最新人员组织概览
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrganizationOverview"
        "404":
          description: 角色不存在
    delete:
      summary: 停用租户业务角色
      description: 仅当没有启用成员使用该角色时允许停用。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 停用成功
        "400":
          description: 角色下仍有启用成员
        "404":
          description: 角色不存在
  /api/admin/tenants/{tenantId}/organization/memberships/{membershipId}/role:
    patch:
      summary: 调整成员角色
      description: 更新成员关系中的角色，后端会校验目标角色归属当前租户且为启用状态。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: membershipId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMembershipRoleRequest"
      responses:
        "200":
          description: 更新成功并返回最新人员组织概览
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrganizationOverview"
        "400":
          description: 角色不可用或不属于当前租户
        "401":
          description: 未登录或令牌无效
        "403":
          description: 当前账号不能管理该租户
        "404":
          description: 成员关系不存在
  /api/admin/tenants/{tenantId}/organization/memberships/{membershipId}/department:
    patch:
      summary: 调整成员部门
      description: 更新成员关系中的部门，可置空。后端会校验目标部门归属当前租户且为启用状态。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: membershipId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMembershipDepartmentRequest"
      responses:
        "200":
          description: 更新成功并返回最新人员组织概览
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrganizationOverview"
        "400":
          description: 部门不可用或不属于当前租户
        "401":
          description: 未登录或令牌无效
        "403":
          description: 当前账号不能管理该租户
        "404":
          description: 成员关系不存在
  /api/admin/tenants/{tenantId}/organization/memberships/{membershipId}/status:
    patch:
      summary: 调整成员状态
      description: 启用或禁用成员关系。禁用不会删除历史关系，但会影响当前租户入口切换和业务可见性。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: membershipId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMembershipStatusRequest"
      responses:
        "200":
          description: 更新成功并返回最新人员组织概览
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrganizationOverview"
        "400":
          description: 成员状态非法
        "401":
          description: 未登录或令牌无效
        "403":
          description: 当前账号不能管理该租户
        "404":
          description: 成员关系不存在
  /api/admin/tenants/{tenantId}/organization/org-roles:
    get:
      summary: 分页获取租户内角色权限
      description: 租户管理页使用的第二层页面权限配置台账。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: sort
          in: query
          schema:
            type: string
            default: updatedAt,desc
      responses:
        "200":
          description: 租户内角色分页
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrgRolePageResponse"
        "401":
          description: 未登录或令牌无效
        "403":
          description: 当前账号不能管理该租户
    post:
      summary: 新增租户内角色权限
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantOrgRoleRequest"
      responses:
        "200":
          description: 新增后的角色权限
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TenantOrgRole"
        "409":
          description: 租户内角色名称重复
  /api/admin/tenants/{tenantId}/organization/resource-options:
    get:
      summary: 获取当前租户可授权资源
      description: 只返回系统管理已启用给当前租户的 MCP、Skill、提示词模板和交付能力，用于租户内资源授权。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 当前租户可授权资源
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/TenantResourceOption"
        "401":
          description: 未登录或令牌无效
        "403":
          description: 当前账号不能管理该租户
  /api/admin/tenants/{tenantId}/organization/page-grants:
    get:
      summary: 获取租户页签授权规则
      description: 返回按角色、部门或人员授权的业务侧页签入口规则。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 页签授权规则列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/PageGrant"
    post:
      summary: 新增租户页签授权
      description: 授权主体支持 role、department、user；页签范围仅限业务工作台、流程设计、能力资产和运行审计。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePageGrantRequest"
      responses:
        "200":
          description: 新增后的页签授权规则
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageGrant"
        "400":
          description: 授权主体或页签不可用
        "409":
          description: 授权规则已存在
  /api/admin/tenants/{tenantId}/organization/page-grants/{grantId}:
    delete:
      summary: 删除租户页签授权
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: grantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 删除成功
        "404":
          description: 授权规则不存在
  /api/admin/tenants/{tenantId}/organization/resource-grants:
    get:
      summary: 获取租户能力授权规则
      description: 返回按角色、部门或人员直接授权的能力资源规则。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 能力授权规则列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ResourceGrant"
    post:
      summary: 新增租户能力授权
      description: 授权主体支持 role、department、user；资源必须是系统管理已启用给当前租户的能力。
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceGrantRequest"
      responses:
        "200":
          description: 新增后的能力授权规则
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ResourceGrant"
        "400":
          description: 授权主体或资源不可用
        "409":
          description: 授权规则已存在
  /api/admin/tenants/{tenantId}/organization/resource-grants/{grantId}:
    delete:
      summary: 删除租户能力授权
      tags:
        - Organization
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: grantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 删除成功
        "404":
          description: 授权规则不存在
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          anyOf:
            - $ref: "#/components/schemas/ApiError"
            - type: "null"
        requestId:
          type: string
      required:
        - success
        - data
        - error
        - requestId
...
    TenantOrganizationOverview:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        tenantName:
          type: string
        tenantCode:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationMember"
        departments:
          type: array
          items:
            $ref: "#/components/schemas/Department"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationRole"
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/UserMembership"
      required:
        - tenantId
        - tenantName
        - tenantCode
        - members
        - departments
        - roles
        - memberships
    CreateMemberRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 100
        password:
          type: string
          minLength: 8
          maxLength: 80
        displayName:
          type: string
          maxLength: 100
        email:
          type: string
          maxLength: 255
        departmentId:
          type: string
          format: uuid
          nullable: true
        roleId:
          type: string
          format: uuid
        spaceCode:
          type: string
          maxLength: 80
      required:
        - username
        - password
        - displayName
        - roleId
    CreateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 160
        parentId:
          type: string
          format: uuid
          nullable: true
        sortOrder:
          type: integer
      required:
        - name
...